FRforms: change from observer to pubsub
For the breedselect-geneselect relationship, a pubsub pattern feels cleaner than the observer pattern that was used previously. Now the selects don't need to hold references to each other and query the document for each other if they're attached/detached in a weird order; they just consult one singular message bus.