fr/data: Gene ancient lists contain indices, not names
Gene objects now provide breed INDICES in their list of available ancient breeds, as opposed to names like they did before. This is mainly for consistency reasons; every other data point that references another data point in fr/data references it either by its index in an array, or by an enum property.

Looks small in the diff, but a big enough change to the actual data that it warrants its own commit.