Restructure genes with SIDs
Genes now use a dictionary object where keys are a breed ID (or "M" for modern) the gene can be applied to, and values are the site ID when the gene is on that breed; the `modern` and `ancient` properties have been removed.

The dictionary stores three important types of information in the least amount of space possible: whether the gene is available on moderns or not, what ancients it's available on if any, and what each in-game on-site IDs is for every breed it can be on. Genes also now have convenience functions for retrieving that info, to mimic the removed properties' functionality.

Partial progress on #1 .