data.js; Generators yield indices, not objects
Generating objects which were effectively a data object except it has an index property tacked on was clunky, hard to document properly, and inconsistent with the fact that the module expects you to be working with indexes the majority of the time.

Generating indexes is cleaner, and if the user needs the actual data objects it's only one more line of code to grab it directly from the relevant array.

BREAKING CHANGE: genesForBreed and colourRange functions in FRjs/data now yield indices on the GENES and COLOURS arrays, respectively, instead of objects.