data.js; minimizing & minor optimizations
Gene creation function now uses a rest parameter for sid key-value pairs and constructs the sids map itself. This is very slightly slower to initialize than the previous way, but it removes ~1.7KB of characters that cannot be minimized away.

Also, most data objects freeze themselves on creation now, rather than being created then having to be iterated over again to get frozen. Extremely minor; but freezing is the most expensive part of the module's start up so I try to cut it down as much as possible.