docs: Swap namepaths in types for short names
Because of how JSDoc handles namepaths for modules, the correct namepath for referencing a module member in documentation is long, hard to read, and incompatible with VSCode's intellisense.

In the interest of keeping the development process smooth and the docs more readable, custom classes/typedefs are referred to by their intellisense compatible short name wherever possible; and to mitigate the loss of those types becoming automatic links to the right section for each typedef/class/etc in the docs, doc comments now include @see tags.